cleanup/reorganize: move CompiledQuery out of 'fes' logic.
This splits the ORM lookup logic (needed for various operators),
from the query building that happens in both fes20 and 'gml' namespaces.
