+++ gisserver/parsers/fes20/sorting.py
@@ -5,6 +5,7 @@ from enum import Enum
 
 from gisserver.exceptions import InvalidParameterValue
 from gisserver.parsers.fes20 import ValueReference
+from gisserver.parsers.query import CompiledQuery
 from gisserver.parsers.xml import NSElement, get_child, split_ns
 
 
@@ -88,14 +89,15 @@ class SortBy:
             props.append(sort_property)
         return cls(sort_properties=props)
 
-    def build_ordering(self, feature_type=None) -> list[str]:
+    def build_ordering(self, compiler: CompiledQuery):
         """Build the ordering for the Django ORM call."""
         ordering = []
         for prop in self.sort_properties:
-            if feature_type is not None:
-                orm_path = prop.value_reference.parse_xpath(feature_type).orm_path
+            if compiler.feature_type is not None:
+                orm_path = prop.value_reference.parse_xpath(compiler.feature_type).orm_path
             else:
                 orm_path = prop.value_reference.xpath.replace("/", "__")
 
             ordering.append(f"{prop.sort_order.value}{orm_path}")
-        return ordering
+
+        compiler.add_ordering(ordering)
